Refactor advertisement service methods to include seller ID as a parameter for enhanced authorization checks

+++ WatchPlatform/WatchCollection.Domain.Services/AdvertisementService.cs
@@ -12,14 +12,14 @@ namespace WatchCollection.Domain.Services;
 
 public class AdvertisementService(IAdvertisementRepository _repository, IWatchValuationHttpClient _watchValuationClient) : IAdvertisementService
 {
-    public async Task<AdvertisementResponseContract> CreateAdvertisement(AdvertisementRequestContract contract)
+    public async Task<AdvertisementResponseContract> CreateAdvertisement(string bidderIdString, AdvertisementRequestContract contract)
     {
         if(contract.Status is AdvertisementStatus.Sold || contract.Status is AdvertisementStatus.Expired)
             throw new InvalidAdvertisementStatusException("Cannot create an advertisement with status Sold or Expired.");
 
         var model = contract.AsModel();
         var advertisementId = Guid.NewGuid();
-        var sellerUserId = Guid.NewGuid(); // This should be retrieved from the authenticated user's context in a real application.
+        var sellerUserId = Guid.TryParse(bidderIdString, out var userIdGuid) ? userIdGuid : throw new Exception("Invalid User ID format.");
         model.AdvertisementId = advertisementId;
         model.SellerUserId = sellerUserId;
         model.ViewCount = 0;
@@ -29,16 +29,31 @@ public class AdvertisementService(IAdvertisementRepository _repository, IWatchVa
         return createdEntity.AsModel().AsContract();
     }
 
-    public async Task DeleteAdvertisement(Guid advertisementId)
+    public async Task DeleteAdvertisement(string sellerIdString, Guid advertisementId)
     {
+        var advertisement = await _repository.GetAdvertisementByIdAsync(advertisementId);
+        if (advertisement is null)
+            throw new AdvertisementNotFoundExceptions(advertisementId, "Advertisement not found");
+        if (advertisement.SellerUserId.ToString() != sellerIdString){
+            // TODO: write override logic for admin users
+            // TODO: write custom exception for unauthorized access
+            throw new UnauthorizedAccessException("User is not authorized to delete this advertisement.");
+        }
         await _repository.DeleteAdvertisementAsync(advertisementId);
     }
 
-    public async Task<AdvertisementResponseContract?> GetAdvertisementById(Guid advertisementId)
+    public async Task<AdvertisementResponseContract?> GetAdvertisementById(string sellerIdString,Guid advertisementId)
     {
         var entity =  await _repository.GetAdvertisementByIdAsync(advertisementId);
         if (entity is null)
             return null;
+        
+        if (entity.SellerUserId.ToString() != sellerIdString){
+            entity.ViewCount += 1;
+            // TODO: Ensure a update view count exists only updating that field
+            await _repository.UpdateAdvertisementAsync(entity);
+        }
+
         return entity.AsModel().AsContract();
     }
 
@@ -48,13 +63,20 @@ public class AdvertisementService(IAdvertisementRepository _repository, IWatchVa
         return entities.Select(e => e.AsModel().AsContract());
     }
 
-    public async Task<AdvertisementResponseContract> UpdateAdvertisement(Guid advertisementId, AdvertisementUpdateRequestContract contract)
+    public async Task<AdvertisementResponseContract> UpdateAdvertisement(Guid advertisementId, string sellerIdString, AdvertisementUpdateRequestContract contract)
     {
         if (contract.Status is AdvertisementStatus.Expired || contract.Status is AdvertisementStatus.Draft)
             throw new InvalidAdvertisementStatusException("Cannot update an advertisement to status Expired or Draft");
+
         var advertisement = await _repository.GetAdvertisementByIdAsync(advertisementId);
         if (advertisement is null)
             throw new AdvertisementNotFoundExceptions(advertisementId, "Advertisement not found");
+
+        if(advertisement.SellerUserId.ToString() != sellerIdString)
+        {
+            // TODO: write override logic for admin users
+            throw new UnauthorizedAccessException("User is not authorized to update this advertisement.");
+        }
         
         var contractWithWatchId = new AdvertisementRequestContract
         {
@@ -66,8 +88,6 @@ public class AdvertisementService(IAdvertisementRepository _repository, IWatchVa
             AllowBids = contract.AllowBids
         };
         var model = contractWithWatchId.AsModel();
-        model.AdvertisementId = advertisementId;
-        model.SellerUserId = advertisement.SellerUserId; // This should be retrieved from the authenticated user's context in a real application.
 
         var entity = model.AsEntity();
